Pack synteny mates as parallel arrays; bp coords as Uint32
- SyntenyFeatureData.mates (object[]) → mateStarts/mateEnds (Uint32Array,
  RPC-transferable) + mateRefNames/mateAssemblyNames (string[]). Drops the
  per-feature object allocation and JSON serialization on the worker boundary
- Drop mate.name field — no adapter sets it; the "Name 2" tooltip line was
  always empty and is removed (the parent feature's Name line covers it)
- Switch starts/ends/mateStarts/mateEnds to Uint32Array — bp coords fit in
  32 bits and this matches the convention used elsewhere in jbrowse. Pixel
  offsets stay Float64 since they accumulate across regions × 1/bpPerPx
- Fix identities type annotation (was Float64Array, was actually Float32Array
  since earlier pass)
- Update consumers: getFeatureAtIndex, syriClassification, getTooltip,
  DiagonalizationProgressDialog

+++ plugins/linear-comparative-view/src/LinearSyntenyRPC/buildSyntenyGeometry.ts
@@ -112,8 +112,8 @@ export function buildSyntenyGeometry({
   strands: Int8Array
   names: string[]
   parsedCigars: number[][]
-  starts: Float64Array
-  ends: Float64Array
+  starts: Uint32Array
+  ends: Uint32Array
   drawCIGAR: boolean
   drawCIGARMatchesOnly: boolean
   drawLocationMarkers: boolean
